+++ src/web/assets/cp/src/js/AssetImageEditor.js
@@ -32,7 +32,6 @@ Craft.AssetImageEditor = Garnish.Modal.extend(
     cropperHandles: null,
     cropperGrid: null,
     croppingShade: null,
-    croppingAreaText: null,
 
     // Image state attributes
     imageStraightenAngle: 0,
@@ -1900,7 +1899,6 @@ Craft.AssetImageEditor = Garnish.Modal.extend(
         this.croppingCanvas.remove(this.cropperHandles);
         this.croppingCanvas.remove(this.cropperGrid);
         this.croppingCanvas.remove(this.croppingRectangle);
-        this.croppingCanvas.remove(this.croppingAreaText);
 
         this.croppingCanvas = null;
         this.renderCropper = null;
@@ -1996,7 +1994,6 @@ Craft.AssetImageEditor = Garnish.Modal.extend(
         this.croppingCanvas.remove(this.cropperHandles);
         this.croppingCanvas.remove(this.cropperGrid);
         this.croppingCanvas.remove(this.croppingRectangle);
-        this.croppingCanvas.remove(this.croppingAreaText);
       }
       this._redrawCropperElements._.lineOptions = {
         strokeWidth: 4,
@@ -2134,30 +2131,9 @@ Craft.AssetImageEditor = Garnish.Modal.extend(
           'rgba(0,0,0,0.5)';
       }
 
-      this.croppingAreaText = new fabric.Textbox(
-        Math.round(this.clipper.width) +
-          ' x ' +
-          Math.round(this.clipper.height),
-        {
-          left: this.croppingRectangle.left,
-          top: this._redrawCropperElements._.cropTextTop,
-          fontSize: 13,
-          fill: 'rgb(200,200,200)',
-          backgroundColor:
-            this._redrawCropperElements._.cropTextBackgroundColor,
-          font: 'Craft',
-          width: 70,
-          height: 15,
-          originX: 'center',
-          originY: 'center',
-          textAlign: 'center',
-        }
-      );
-
       this.croppingCanvas.add(this.cropperHandles);
       this.croppingCanvas.add(this.cropperGrid);
       this.croppingCanvas.add(this.croppingRectangle);
-      this.croppingCanvas.add(this.croppingAreaText);
     },
 
     /**